b_complete
b_complete is the button that will copy the code to the user's clipboard while also displaying a message box with the code. Rishit, add it so that it will show up when I click b_encode (I had to utilize b_check to test the code I used for the complete definition).